refactor(template): Extract TEMPLATE_CONTENT to dedicated utility
